flot double solved

+++ src/chatterbox/tts.py
@@ -227,7 +227,7 @@ class ChatterboxTTS:
             self.conds.t3 = T3Cond(
                 speaker_emb=_cond.speaker_emb,
                 cond_prompt_speech_tokens=_cond.cond_prompt_speech_tokens,
-                emotion_adv=exaggeration * torch.ones(1, 1, 1),
+                emotion_adv=exaggeration * torch.ones(1, 1, 1, dtype=torch.float32),
             ).to(device=self.device)
 
         # Norm and tokenize text
